requirements.txt
pyspark==3.4.0
delta-spark==2.4.0
pytest==7.3.1
